requirements.txt
# v0.2: local experiments (JSON/JSONL + optional schema validation + charts + env)
jsonschema>=4.0,<5
python-dotenv>=1.0,<2
matplotlib>=3.8,<4
